+++ src/extension.ts
@@ -1,18 +1,9 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 import { runTestsInFile, runTestsInFileDebug } from './commands';
-import {
-	findEndTestTitle,
-	findTestTitleOpeningCharacter,
-	getTestTitles,
-} from './matcher';
+import { getTestTitles } from './matcher';
+import { Configuration } from './configuration';
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
 export function activate( context: vscode.ExtensionContext ) {
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
 	console.log(
 		'Congratulations, your extension "vsc-ava-test-runner" is now active!'
 	);
@@ -47,16 +38,24 @@ class AvaCodelens implements vscode.CodeLensProvider {
 		const text = document.getText();
 		const results = getTestTitles( text );
 
-		const codeLenses = results.map( ( [ title, match ] ) => {
+		const config = Configuration.load();
+
+		const codeLenses = results.flatMap( ( [ title, match ] ) => {
 			const line = document.lineAt( document.positionAt( match ).line );
 			const position = new vscode.Position( line.lineNumber, 0 );
 			const range = document.getWordRangeAtPosition( position );
 
-			// Need a range to apply CodeLens
-			const command = this.createRunTestCaseCommand( title ); // Anything for now
-			// If title and match were found correctly
-			// It the range have to be valid
-			return new vscode.CodeLens( range!, command );
+			const lenses = [];
+
+			const runCommand = this.createRunTestCaseCommand( title );
+			lenses.push( new vscode.CodeLens( range!, runCommand ) );
+
+			if( config.areExperimentalsEanbled() ){
+				const runByLineCommand = this.createRunTestByLineCommand( line.lineNumber + 1 );
+				lenses.push( new vscode.CodeLens( range!, runByLineCommand ) );
+			}
+
+			return lenses;
 		} );
 
 		return codeLenses;
@@ -69,6 +68,16 @@ class AvaCodelens implements vscode.CodeLensProvider {
 			// https://github.com/avajs/ava/blob/main/docs/05-command-line.md#running-tests-with-matching-titles
 			// * might be problematic it test title
 			arguments: [ ` --match='${testCaseTitle}'` ],
+		};
+
+		return command;
+	}
+
+	createRunTestByLineCommand( lineNumber: number ) {
+		const command = {
+			title: `Run [L:${lineNumber}]`,
+			command: 'vsc-ava-test-runner.runTestsInFile',
+			arguments: [ `:${lineNumber}` ],
 		};
 
 		return command;