+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
-## [0.0.1]
+
+## [0.0.2]
+### Added
+- Add extension setting: `ava-runner.experimentalEnabled`
+- Add experimental feature: run test case via [line number](https://github.com/avajs/ava/blob/main/docs/05-command-line.md#running-tests-at-specific-line-numbers). Use the `ava-runner.experimentalEnabled` config option to enable that feature
+
+## [0.0.1] (22.06.2022)
 
 Initial release.
 ### Added